Module aws_sdk_sms::model

Modules

app_summary

See AppSummary

app_validation_configuration

See AppValidationConfiguration

app_validation_output

See AppValidationOutput

connector

See Connector

launch_details

See LaunchDetails

notification_context

See NotificationContext

replication_job

See ReplicationJob

replication_run

See ReplicationRun

replication_run_stage_details

See ReplicationRunStageDetails

s3_location

See S3Location

server

See Server

server_group

See ServerGroup

server_group_launch_configuration

See ServerGroupLaunchConfiguration

server_group_replication_configuration

See ServerGroupReplicationConfiguration

server_group_validation_configuration

See ServerGroupValidationConfiguration

server_launch_configuration

See ServerLaunchConfiguration

server_replication_configuration

See ServerReplicationConfiguration

server_replication_parameters

See ServerReplicationParameters

server_validation_configuration

See ServerValidationConfiguration

server_validation_output

See ServerValidationOutput

source

See Source

ssm_output

See SsmOutput

ssm_validation_parameters

See SsmValidationParameters

tag

See Tag

user_data

See UserData

user_data_validation_parameters

See UserDataValidationParameters

validation_output

See ValidationOutput

vm_server

See VmServer

vm_server_address

See VmServerAddress

Structs

AppSummary

Information about the application.

AppValidationConfiguration

Configuration for validating an application.

AppValidationOutput

Output from validating an application.

Connector

Represents a connector.

LaunchDetails

Details about the latest launch of an application.

NotificationContext

Contains the status of validating an application.

ReplicationJob

Represents a replication job.

ReplicationRun

Represents a replication run.

ReplicationRunStageDetails

Details of the current stage of a replication run.

S3Location

Location of an Amazon S3 object.

Server

Represents a server.

ServerGroup

Logical grouping of servers.

ServerGroupLaunchConfiguration

Launch configuration for a server group.

ServerGroupReplicationConfiguration

Replication configuration for a server group.

ServerGroupValidationConfiguration

Configuration for validating an instance.

ServerLaunchConfiguration

Launch configuration for a server.

ServerReplicationConfiguration

Replication configuration of a server.

ServerReplicationParameters

The replication parameters for replicating a server.

ServerValidationConfiguration

Configuration for validating an instance.

ServerValidationOutput

Contains output from validating an instance.

Source

Contains the location of a validation script.

SsmOutput

Contains the location of validation output.

SsmValidationParameters

Contains validation parameters.

Tag

Key/value pair that can be assigned to an application.

UserData

A script that runs on first launch of an Amazon EC2 instance. Used for configuring the

UserDataValidationParameters

Contains validation parameters.

ValidationOutput

Contains validation output.

VmServer

Represents a VM server.

VmServerAddress

Represents a VM server location.

Enums

AppLaunchConfigurationStatus

AppLaunchStatus

AppReplicationConfigurationStatus

AppReplicationStatus

AppStatus

AppValidationStrategy

ConnectorCapability

ConnectorStatus

LicenseType

OutputFormat

ReplicationJobState

ReplicationRunState

ReplicationRunType

ScriptType

ServerCatalogStatus

ServerType

ServerValidationStrategy

ValidationStatus

VmManagerType