Module aws_sdk_sms::operation

Structs

CreateApp

Creates an application. An application consists of one or more server groups. Each

CreateReplicationJob

Creates a replication job. The replication job schedules periodic replication runs

DeleteApp

Deletes the specified application. Optionally deletes the launched stack associated with

DeleteAppLaunchConfiguration

Deletes the launch configuration for the specified application.

DeleteAppReplicationConfiguration

Deletes the replication configuration for the specified application.

DeleteAppValidationConfiguration

Deletes the validation configuration for the specified application.

DeleteReplicationJob

Deletes the specified replication job.

DeleteServerCatalog

Deletes all servers from your server catalog.

DisassociateConnector

Disassociates the specified connector from AWS SMS.

GenerateChangeSet

Generates a target change set for a currently launched stack and writes it to an Amazon S3

GenerateTemplate

Generates an AWS CloudFormation template based on the current launch configuration and writes it to

GetApp

Retrieve information about the specified application.

GetAppLaunchConfiguration

Retrieves the application launch configuration associated with the specified application.

GetAppReplicationConfiguration

Retrieves the application replication configuration associated with the specified

GetAppValidationConfiguration

Retrieves information about a configuration for validating an application.

GetAppValidationOutput

Retrieves output from validating an application.

GetConnectors

Describes the connectors registered with the AWS SMS.

GetReplicationJobs

Describes the specified replication job or all of your replication jobs.

GetReplicationRuns

Describes the replication runs for the specified replication job.

GetServers

Describes the servers in your server catalog.

ImportAppCatalog

Allows application import from AWS Migration Hub.

ImportServerCatalog

Gathers a complete list of on-premises servers. Connectors must be installed and

LaunchApp

Launches the specified application as a stack in AWS CloudFormation.

ListApps

Retrieves summaries for all applications.

NotifyAppValidationOutput

Provides information to AWS SMS about whether application validation is successful.

PutAppLaunchConfiguration

Creates or updates the launch configuration for the specified application.

PutAppReplicationConfiguration

Creates or updates the replication configuration for the specified application.

PutAppValidationConfiguration

Creates or updates a validation configuration for the specified application.

StartAppReplication

Starts replicating the specified application by creating replication jobs for each server in the

StartOnDemandAppReplication

Starts an on-demand replication run for the specified application.

StartOnDemandReplicationRun

Starts an on-demand replication run for the specified replication job. This

StopAppReplication

Stops replicating the specified application by deleting the replication job for each server in

TerminateApp

Terminates the stack for the specified application.

UpdateApp

Updates the specified application.

UpdateReplicationJob

Updates the specified settings for the specified replication job.